refactor(config): OAuth 관련 API 설명을 적절한 위치로 이동시킨다
